fix: moved MEETING_KEY from collection.ts to meetings.types.ts
