+++ packages/@webex/plugin-meetings/src/meetings/collection.ts
@@ -1,18 +1,7 @@
 import {find} from 'lodash';
-import {Enum} from '../constants';
 
 import Collection from '../common/collection';
-
-export const MEETING_KEY = {
-  CONVERSATION_URL: 'conversationUrl',
-  SIP_URI: 'sipUri',
-  LOCUS_URL: 'locusUrl',
-  MEETINGNUMBER: 'meetingNumber',
-  CORRELATION_ID: 'correlationId',
-} as const;
-
-export type MEETING_KEY = Enum<typeof MEETING_KEY>;
-
+import {MEETING_KEY} from './meetings.types';
 /**
  * @export
  * @class MeetingCollection
